requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
defusedxml==0.7.1
docopt==0.6.2
fastjsonschema==2.20.0
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.4.0
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.6
importlib_resources==6.4.5
ipython==8.12.3
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab_pygments==0.3.0
keras==2.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.6.3
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.21.0
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
pandas==1.5.3
pandocfilters==1.5.1
Pillow==10.1.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.0
six==1.15.0
soupsieve==2.6
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
threadpoolctl==3.2.0
tinycss2==1.3.0
typing-extensions==3.7.4.3
urllib3==2.2.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.12.1
yarg==0.1.9

